float conversion not needed anymore

+++ src/activity_handlers.cpp
@@ -1588,7 +1588,7 @@ void activity_handlers::fill_liquid_do_turn( player_activity *act, player *p )
         const int original_charges = liquid.charges;
         if( liquid.has_temperature() && liquid.specific_energy < 0 ) {
             liquid.set_item_temperature( temp_to_kelvin( std::max( get_weather().get_temperature( p->pos() ),
-                                         static_cast<double>( temperatures::cold ) ) ) );
+                                         temperatures::cold ) ) );
         }
 
         // 2. Transfer charges.
